% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{angular_difference}
\alias{angular_difference}
\title{Signed angular difference}
\usage{
angular_difference(x, y, period = 2 * pi)
}
\arguments{
\item{x, y}{Numeric vectors of angles.}

\item{period}{Positive numeric period.}
}
\value{
A numeric vector following R recycling rules.
}
\description{
Computes the signed difference \code{x - y} on a periodic scale. With the default
period, values are returned in \verb{[-pi, pi)}.
}
\examples{
angular_difference(0, 3 * pi / 2)
}
\seealso{
Other angle utilities: 
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
