% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{angular_distance}
\alias{angular_distance}
\title{Circular angular distance}
\usage{
angular_distance(x, y, period = 2 * pi)
}
\arguments{
\item{x, y}{Numeric vectors of angles.}

\item{period}{Positive numeric period.}
}
\value{
A non-negative numeric vector.
}
\description{
Computes the non-negative angular distance between \code{x} and \code{y}. With the
default period, values are returned in \verb{[0, pi]}.
}
\examples{
angular_distance(0, 3 * pi / 2)
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
