% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot_circular}
\alias{autoplot_circular}
\title{Autoplot circular data}
\usage{
autoplot_circular(
  theta,
  bins = 24,
  density = TRUE,
  mean = TRUE,
  axial = FALSE,
  ...
)
}
\arguments{
\item{theta}{Numeric vector of angles in radians.}

\item{bins}{Number of rose diagram bins.}

\item{density}{Should a circular density estimate be added?}

\item{mean}{Should the mean direction be added?}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{...}{Additional arguments currently ignored.}
}
\value{
A ggplot object.
}
\description{
Creates a quick diagnostic plot for a numeric vector of circular angles.
}
\examples{
autoplot_circular(wind_directions$direction)
}
\concept{autoplot helpers}
