% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-draws.R
\name{autoplot_circular_draws}
\alias{autoplot_circular_draws}
\title{Autoplot circular posterior draws}
\usage{
autoplot_circular_draws(
  draws,
  variables = NULL,
  type = c("density", "interval"),
  axial = FALSE,
  ...
)
}
\arguments{
\item{draws}{Circular draws or posterior draws.}

\item{variables}{Optional variables to plot.}

\item{type}{Plot type.}

\item{axial}{Should draws be treated as axial, modulo \code{pi}?}

\item{...}{Additional arguments passed to \code{\link[=as_circular_draws]{as_circular_draws()}} when needed.}
}
\value{
A ggplot object.
}
\description{
Autoplot circular posterior draws
}
\seealso{
Other posterior helpers: 
\code{\link{as_circular_draws}()},
\code{\link{summarise_circular_draws}()}
}
\concept{posterior helpers}
