% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{axial_orientations}
\alias{axial_orientations}
\title{Simulated axial orientations}
\format{
A tibble with 300 rows and 3 variables:
\describe{
\item{sample}{Sample identifier.}
\item{orientation}{Axial orientation in radians, modulo \code{pi}.}
\item{group}{Group label.}
}
}
\source{
Simulated for package examples.
}
\usage{
axial_orientations
}
\description{
Simulated axial orientation data, such as fiber or fault orientations, stored
modulo \code{pi}.
}
\keyword{datasets}
