% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical.R
\name{cartesian_to_spherical}
\alias{cartesian_to_spherical}
\title{Convert Cartesian coordinates to spherical coordinates}
\usage{
cartesian_to_spherical(
  x,
  y,
  z,
  convention = c("azimuth_colatitude", "azimuth_elevation")
)
}
\arguments{
\item{x, y, z}{Cartesian coordinates.}

\item{convention}{Output convention for \code{phi}.}
}
\value{
A tibble with \code{theta}, \code{phi} and \code{radius}.
}
\description{
Convert Cartesian coordinates to spherical coordinates
}
\seealso{
Other spherical helpers: 
\code{\link{spherical_summary}()},
\code{\link{spherical_to_cartesian}()}
}
\concept{spherical helpers}
