% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{check_angle}
\alias{check_angle}
\title{Check an angle vector}
\usage{
check_angle(x, allow_na = TRUE)
}
\arguments{
\item{x}{Object to check.}

\item{allow_na}{Should missing values be allowed?}
}
\value{
Invisibly returns \code{x} if the check succeeds.
}
\description{
Check an angle vector
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
