% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-circular.R
\name{circular_sd}
\alias{circular_sd}
\title{Circular standard deviation}
\usage{
circular_sd(x, axial = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be removed?}
}
\value{
Circular standard deviation in radians.
}
\description{
Uses the common descriptive statistic \code{sqrt(-2 * log(Rbar))}.
}
\seealso{
Other circular summaries: 
\code{\link{circular_mean_ci}()},
\code{\link{circular_summary}()},
\code{\link{circular_variance}()},
\code{\link{estimate_kappa}()},
\code{\link{mean_direction}()},
\code{\link{mean_resultant_length}()},
\code{\link{resultant_length}()}
}
\concept{circular summaries}
