% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-circular.R
\name{circular_variance}
\alias{circular_variance}
\title{Circular variance}
\usage{
circular_variance(x, axial = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be removed?}
}
\value{
The circular variance \code{1 - Rbar}.
}
\description{
Circular variance
}
\seealso{
Other circular summaries: 
\code{\link{circular_mean_ci}()},
\code{\link{circular_sd}()},
\code{\link{circular_summary}()},
\code{\link{estimate_kappa}()},
\code{\link{mean_direction}()},
\code{\link{mean_resultant_length}()},
\code{\link{resultant_length}()}
}
\concept{circular summaries}
