% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{compass_to_rad}
\alias{compass_to_rad}
\title{Convert compass labels to radians}
\usage{
compass_to_rad(x)
}
\arguments{
\item{x}{Character vector of compass labels.}
}
\value{
Numeric vector of bearing angles in radians.
}
\description{
Converts the eight standard compass labels \code{N}, \code{NE}, \code{E}, \code{SE}, \code{S}, \code{SW},
\code{W} and \code{NW} to bearing angles in radians, where zero is north and angles
increase clockwise.
}
\examples{
compass_to_rad(c("N", "E", "S", "W"))
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
