% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement-features.R
\name{compute_bearing}
\alias{compute_bearing}
\title{Compute bearings}
\usage{
compute_bearing(x, y, angle_convention = c("mathematical", "bearing"))
}
\arguments{
\item{x, y}{Numeric coordinate vectors.}

\item{angle_convention}{Angle convention. \code{"mathematical"} means zero is east
and angles increase counterclockwise. \code{"bearing"} means zero is north and
angles increase clockwise.}
}
\value{
Numeric vector of bearings in radians. The first value is \code{NA}.
}
\description{
Compute bearings
}
\seealso{
Other movement helpers: 
\code{\link{as_step_data}()},
\code{\link{augment_momentuHMM_angles}()},
\code{\link{compute_step_length}()},
\code{\link{compute_turn_angle}()},
\code{\link{geom_circular_point}()},
\code{\link{geom_direction_arrow}()},
\code{\link{mutate_directional_features}()},
\code{\link{plot_state_angles}()}
}
\concept{movement helpers}
