% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement-features.R
\name{compute_step_length}
\alias{compute_step_length}
\title{Compute step lengths}
\usage{
compute_step_length(x, y)
}
\arguments{
\item{x, y}{Numeric coordinate vectors.}
}
\value{
Numeric vector of Euclidean step lengths. The first value is \code{NA}.
}
\description{
Compute step lengths
}
\seealso{
Other movement helpers: 
\code{\link{as_step_data}()},
\code{\link{augment_momentuHMM_angles}()},
\code{\link{compute_bearing}()},
\code{\link{compute_turn_angle}()},
\code{\link{geom_circular_point}()},
\code{\link{geom_direction_arrow}()},
\code{\link{mutate_directional_features}()},
\code{\link{plot_state_angles}()}
}
\concept{movement helpers}
