% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement-features.R
\name{compute_turn_angle}
\alias{compute_turn_angle}
\title{Compute turn angles}
\usage{
compute_turn_angle(bearing, period = 2 * pi)
}
\arguments{
\item{bearing}{Numeric vector of bearings in radians.}

\item{period}{Angular period.}
}
\value{
Numeric vector of signed turn angles. The first value is \code{NA}.
}
\description{
Compute turn angles
}
\seealso{
Other movement helpers: 
\code{\link{as_step_data}()},
\code{\link{augment_momentuHMM_angles}()},
\code{\link{compute_bearing}()},
\code{\link{compute_step_length}()},
\code{\link{geom_circular_point}()},
\code{\link{geom_direction_arrow}()},
\code{\link{mutate_directional_features}()},
\code{\link{plot_state_angles}()}
}
\concept{movement helpers}
