% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert degrees to radians}
\usage{
deg_to_rad(x)
}
\arguments{
\item{x}{Numeric vector in degrees.}
}
\value{
Numeric vector in radians.
}
\description{
Convert degrees to radians
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
