% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-circular.R
\name{estimate_kappa}
\alias{estimate_kappa}
\title{Estimate von Mises concentration}
\usage{
estimate_kappa(x, axial = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be removed?}
}
\value{
Estimated concentration parameter \code{kappa}.
}
\description{
Estimates the von Mises concentration parameter from the mean resultant
length using the standard piecewise approximation described by Fisher (1993).
}
\references{
Fisher, N. I. (1993). \emph{Statistical Analysis of Circular Data}.
Cambridge University Press.
}
\seealso{
Other circular summaries: 
\code{\link{circular_mean_ci}()},
\code{\link{circular_sd}()},
\code{\link{circular_summary}()},
\code{\link{circular_variance}()},
\code{\link{mean_direction}()},
\code{\link{mean_resultant_length}()},
\code{\link{resultant_length}()}
}
\concept{circular summaries}
