% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional-layers.R
\name{geom_phase_loom}
\alias{geom_phase_loom}
\title{Add a phase loom layer}
\usage{
geom_phase_loom(
  mapping = NULL,
  data = NULL,
  position = "identity",
  n_sectors = 48,
  n_x_bins = 24,
  min_mass = 0.002,
  max_flows = 180,
  mass_type = c("joint", "conditional"),
  curvature = 0.24,
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, position, show.legend, inherit.aes}{Standard ggplot2 layer
arguments.}

\item{n_sectors}{Number of angular sectors.}

\item{n_x_bins}{Number of bins for the linear response.}

\item{min_mass}{Minimum mass to draw.}

\item{max_flows}{Maximum number of flows to draw. Use \code{NULL} to draw all.}

\item{mass_type}{Either \code{"joint"} or \code{"conditional"}.}

\item{curvature}{Curvature passed to \code{ggplot2::geom_curve()}.}

\item{...}{Additional arguments passed to the geom.}
}
\value{
A ggplot2 layer.
}
\description{
Adds mass-flow curves for circular-linear data. Use this layer with
\code{ggplot2::ggplot(data, ggplot2::aes(x = theta, y = response))}.
}
\examples{
dat <- simulate_cylindrical(n = 80, seed = 1)
ggplot2::ggplot(dat, ggplot2::aes(x = theta, y = x)) +
  geom_phase_loom(n_sectors = 12, n_x_bins = 8, min_mass = 0)
}
\seealso{
Other cylindrical dependence layers: 
\code{\link{stat_circular_topography}()},
\code{\link{stat_statistical_orbit}()},
\code{\link{stat_statistical_orbit_ribbon}()}
}
\concept{cylindrical dependence layers}
