% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional-layers.R
\name{geom_toroidal_flow}
\alias{geom_toroidal_flow}
\title{Add a toroidal flow layer}
\usage{
geom_toroidal_flow(
  mapping = NULL,
  data = NULL,
  position = "identity",
  n_sectors = 32,
  min_mass = 0.002,
  mass_type = c("joint", "conditional"),
  curvature = 0.28,
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, position, show.legend, inherit.aes}{Standard ggplot2 layer
arguments.}

\item{n_sectors}{Number of angular sectors for both variables.}

\item{min_mass}{Minimum mass to draw.}

\item{mass_type}{Either \code{"joint"} or \code{"conditional"}.}

\item{curvature}{Curvature passed to \code{ggplot2::geom_curve()}.}

\item{...}{Additional arguments passed to the geom.}
}
\value{
A ggplot2 layer.
}
\description{
Adds rectangular flow curves from \code{theta} sectors to \code{phi} sectors. Use with
\code{aes(x = theta, y = phi)}.
}
\examples{
dat <- simulate_toroidal(n = 80, seed = 1)
ggplot2::ggplot(dat, ggplot2::aes(x = theta, y = phi)) +
  geom_toroidal_flow(n_sectors = 12, min_mass = 0)
}
\seealso{
Other toroidal dependence layers: 
\code{\link{stat_toroidal_ridge}()},
\code{\link{stat_toroidal_topography}()}
}
\concept{toroidal dependence layers}
