% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{hour_to_rad}
\alias{hour_to_rad}
\title{Convert hours to radians}
\usage{
hour_to_rad(x)
}
\arguments{
\item{x}{Numeric vector in hours on a 24-hour clock.}
}
\value{
Numeric vector in radians.
}
\description{
Convert hours to radians
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
