% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hourly_activity}
\alias{hourly_activity}
\title{Simulated hourly activity}
\format{
A tibble with 240 rows and 5 variables:
\describe{
\item{id}{Individual identifier.}
\item{hour}{Hour of day.}
\item{angle}{Hour converted to radians.}
\item{activity}{Activity level.}
\item{group}{Group label.}
}
}
\source{
Simulated for package examples.
}
\usage{
hourly_activity
}
\description{
A reproducible simulated dataset of hourly activity converted to circular
angles.
}
\keyword{datasets}
