% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{is_angle}
\alias{is_angle}
\title{Test whether an object can represent angles}
\usage{
is_angle(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} when \code{x} is numeric and contains only finite values or \code{NA}.
}
\description{
Test whether an object can represent angles
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
