% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-angles.R
\name{plot_state_angles}
\alias{plot_state_angles}
\title{Plot angular distributions by state}
\usage{
plot_state_angles(
  data,
  angle,
  state,
  type = c("rose", "density", "mean"),
  bins = 24,
  axial = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{angle}{Angle column.}

\item{state}{State or group column.}

\item{type}{Plot type.}

\item{bins}{Number of bins for rose diagrams.}

\item{axial}{Should data be treated as axial, modulo \code{pi}?}
}
\value{
A ggplot object.
}
\description{
Convenience function for visualizing angles by observed or inferred states.
}
\seealso{
Other movement helpers: 
\code{\link{as_step_data}()},
\code{\link{augment_momentuHMM_angles}()},
\code{\link{compute_bearing}()},
\code{\link{compute_step_length}()},
\code{\link{compute_turn_angle}()},
\code{\link{geom_circular_point}()},
\code{\link{geom_direction_arrow}()},
\code{\link{mutate_directional_features}()}
}
\concept{movement helpers}
