% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{rad_to_compass}
\alias{rad_to_compass}
\title{Convert radians to compass labels}
\usage{
rad_to_compass(x, labels = c("N", "NE", "E", "SE", "S", "SW", "W", "NW"))
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{labels}{Character vector of equally spaced labels.}
}
\value{
Character vector of labels.
}
\description{
Converts angles to the nearest label among \code{labels}. Angles are interpreted
as bearings by default: zero is north and angles increase clockwise.
}
\examples{
rad_to_compass(c(0, pi / 2, pi))
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_deg}()},
\code{\link{rad_to_hour}()}
}
\concept{angle utilities}
