% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-angle.R
\name{rad_to_hour}
\alias{rad_to_hour}
\title{Convert radians to hours}
\usage{
rad_to_hour(x)
}
\arguments{
\item{x}{Numeric vector in radians.}
}
\value{
Numeric vector in hours on a 24-hour clock.
}
\description{
Convert radians to hours
}
\seealso{
Other angle utilities: 
\code{\link{angular_difference}()},
\code{\link{angular_distance}()},
\code{\link{check_angle}()},
\code{\link{compass_to_rad}()},
\code{\link{deg_to_rad}()},
\code{\link{hour_to_rad}()},
\code{\link{is_angle}()},
\code{\link{normalize_angle}()},
\code{\link{rad_to_compass}()},
\code{\link{rad_to_deg}()}
}
\concept{angle utilities}
