% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular-tests.R
\name{rayleigh_test}
\alias{rayleigh_test}
\title{Rayleigh test for circular uniformity}
\usage{
rayleigh_test(x, axial = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{axial}{Should data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be removed?}
}
\value{
An object of class \code{htest}.
}
\description{
Performs the one-sample Rayleigh test for non-uniformity. The returned object
follows the base \code{htest} structure.
}
\seealso{
Other circular tests: 
\code{\link{stat_circular_test}()},
\code{\link{watson_williams_test}()}
}
\concept{circular tests}
