% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-circular.R
\name{resultant_length}
\alias{resultant_length}
\title{Resultant length}
\usage{
resultant_length(x, axial = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of angles in radians.}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be removed?}
}
\value{
The sample resultant length \code{R}.
}
\description{
Resultant length
}
\seealso{
Other circular summaries: 
\code{\link{circular_mean_ci}()},
\code{\link{circular_sd}()},
\code{\link{circular_summary}()},
\code{\link{circular_variance}()},
\code{\link{estimate_kappa}()},
\code{\link{mean_direction}()},
\code{\link{mean_resultant_length}()}
}
\concept{circular summaries}
