% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-circular.R
\name{scale_x_circular_radians}
\alias{scale_x_circular_radians}
\alias{scale_x_circular_degrees}
\alias{scale_x_circular_hours}
\alias{scale_x_circular_compass}
\title{Circular x scales}
\usage{
scale_x_circular_radians(
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = c(0, 2 * pi),
  ...
)

scale_x_circular_degrees(
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = c(0, 2 * pi),
  ...
)

scale_x_circular_hours(
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = c(0, 2 * pi),
  ...
)

scale_x_circular_compass(
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = c(0, 2 * pi),
  ...
)
}
\arguments{
\item{breaks}{Break positions in radians.}

\item{labels}{Break labels.}

\item{limits}{Scale limits in radians.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_x_continuous()}.}
}
\value{
A ggplot2 scale.
}
\description{
These scales label angular x axes in radians, degrees, hours or compass
directions.
}
\examples{
scale_x_circular_radians()
}
\seealso{
Other circular scales: 
\code{\link{coord_circular}()}
}
\concept{circular scales}
