% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherical.R
\name{spherical_to_cartesian}
\alias{spherical_to_cartesian}
\title{Convert spherical coordinates to Cartesian coordinates}
\usage{
spherical_to_cartesian(
  theta,
  phi,
  radius = 1,
  convention = c("azimuth_colatitude", "azimuth_elevation")
)
}
\arguments{
\item{theta}{Azimuth angle in radians.}

\item{phi}{Colatitude or elevation angle in radians.}

\item{radius}{Radius.}

\item{convention}{Interpretation of \code{phi}.}
}
\value{
A tibble with \code{x}, \code{y} and \code{z}.
}
\description{
Convert spherical coordinates to Cartesian coordinates
}
\seealso{
Other spherical helpers: 
\code{\link{cartesian_to_spherical}()},
\code{\link{spherical_summary}()}
}
\concept{spherical helpers}
