% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular-tests.R
\name{stat_circular_test}
\alias{stat_circular_test}
\title{Annotate circular tests}
\usage{
stat_circular_test(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  ...,
  test = c("rayleigh", "watson_williams"),
  x = 0,
  y = 1,
  digits = 3,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, geom, position, show.legend, inherit.aes}{Standard ggplot2
layer arguments.}

\item{...}{Additional arguments passed to the text geom.}

\item{test}{Test to compute.}

\item{x, y}{Text position.}

\item{digits}{Number of digits used for p-value formatting.}

\item{na.rm}{Should missing values be removed?}
}
\value{
A ggplot2 layer.
}
\description{
Adds a text annotation with the p-value from a Rayleigh or Watson-Williams
test.
}
\seealso{
Other circular tests: 
\code{\link{rayleigh_test}()},
\code{\link{watson_williams_test}()}
}
\concept{circular tests}
