% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-density-circle.R
\name{stat_density_circle}
\alias{stat_density_circle}
\title{Smoothed circular density estimates}
\usage{
stat_density_circle(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "stack",
  ...,
  bw = "nrd0",
  adjust = 1,
  kernel = "vonmises",
  n = 512,
  trim = FALSE,
  na.rm = FALSE,
  from = circular(0),
  to = circular(2 * pi),
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Computes density estimates from circular data, when values mapped
to _x_ are positions on a circle, e.g., expressed as angles.
}
