% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional-layers.R
\name{stat_statistical_orbit}
\alias{stat_statistical_orbit}
\title{Add a statistical orbit line}
\usage{
stat_statistical_orbit(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  n_theta = 181,
  kappa = 20,
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, geom, position, show.legend, inherit.aes}{Standard ggplot2
layer arguments.}

\item{n_theta}{Number of angular grid points.}

\item{kappa}{Circular concentration for local weighting.}

\item{...}{Additional arguments passed to the geom.}
}
\value{
A ggplot2 layer.
}
\description{
Add a statistical orbit line
}
\examples{
dat <- simulate_cylindrical(n = 80, seed = 1)
ggplot2::ggplot(dat, ggplot2::aes(x = theta, y = x)) +
  stat_statistical_orbit(n_theta = 24)
}
\seealso{
Other cylindrical dependence layers: 
\code{\link{geom_phase_loom}()},
\code{\link{stat_circular_topography}()},
\code{\link{stat_statistical_orbit_ribbon}()}
}
\concept{cylindrical dependence layers}
