% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-vonmises.R
\name{stat_vonmises}
\alias{stat_vonmises}
\alias{stat_wrapped_normal}
\alias{stat_uniform_circular}
\title{Theoretical von Mises density}
\usage{
stat_vonmises(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  mu = 0,
  kappa = 1,
  n = 512,
  axial = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)

stat_wrapped_normal(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  mu = 0,
  sigma = 1,
  terms = 5,
  n = 512,
  axial = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)

stat_uniform_circular(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  n = 512,
  axial = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping, data, geom, position, show.legend, inherit.aes}{Standard ggplot2
layer arguments.}

\item{...}{Additional arguments passed to the layer.}

\item{mu}{Mean direction in radians.}

\item{kappa}{Non-negative concentration parameter.}

\item{n}{Number of grid points.}

\item{axial}{Should the density be drawn over an axial period of \code{pi}?}

\item{na.rm}{Included for ggplot2 layer compatibility.}

\item{sigma}{Standard deviation of the wrapped normal distribution.}

\item{terms}{Number of wrapping terms on each side of the origin.}
}
\value{
A ggplot2 layer.
}
\description{
Adds a theoretical von Mises density to a circular plot.
}
\examples{
ggplot2::ggplot(wind_directions, ggplot2::aes(x = direction)) +
  geom_rose(ggplot2::aes(y = ggplot2::after_stat(density))) +
  stat_vonmises(mu = pi / 2, kappa = 3)
}
\seealso{
Other circular distributions: 
\code{\link{fit_vonmises_mixture}()},
\code{\link{stat_vonmises_fit}()},
\code{\link{stat_vonmises_mixture}()}
}
\concept{circular distributions}
