% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-vonmises-fit.R
\name{stat_vonmises_fit}
\alias{stat_vonmises_fit}
\title{Fitted von Mises density}
\usage{
stat_vonmises_fit(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  ...,
  n = 512,
  axial = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, geom, position, show.legend, inherit.aes}{Standard ggplot2
layer arguments.}

\item{...}{Additional arguments passed to the layer.}

\item{n}{Number of grid points.}

\item{axial}{Should the data be treated as axial, modulo \code{pi}?}

\item{na.rm}{Should missing values be silently removed?}
}
\value{
A ggplot2 layer.
}
\description{
Estimates \code{mu} with \code{\link[=mean_direction]{mean_direction()}} and \code{kappa} with \code{\link[=estimate_kappa]{estimate_kappa()}},
then draws the fitted von Mises density.
}
\examples{
ggplot2::ggplot(wind_directions, ggplot2::aes(x = direction)) +
  geom_rose(ggplot2::aes(y = ggplot2::after_stat(density))) +
  stat_vonmises_fit()
}
\seealso{
Other circular distributions: 
\code{\link{fit_vonmises_mixture}()},
\code{\link{stat_vonmises}()},
\code{\link{stat_vonmises_mixture}()}
}
\concept{circular distributions}
