% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-draws.R
\name{summarise_circular_draws}
\alias{summarise_circular_draws}
\title{Summarize circular posterior draws}
\usage{
summarise_circular_draws(
  draws,
  variables = NULL,
  level = 0.95,
  axial = FALSE,
  ...
)
}
\arguments{
\item{draws}{Circular draws from \code{\link[=as_circular_draws]{as_circular_draws()}} or any object accepted
by \code{posterior::as_draws_df()}.}

\item{variables}{Optional variables to summarize.}

\item{level}{Credible interval level.}

\item{axial}{Should draws be treated as axial, modulo \code{pi}?}

\item{...}{Additional arguments passed to \code{\link[=as_circular_draws]{as_circular_draws()}} when needed.}
}
\value{
A tibble with posterior circular summaries.
}
\description{
Summarize circular posterior draws
}
\seealso{
Other posterior helpers: 
\code{\link{as_circular_draws}()},
\code{\link{autoplot_circular_draws}()}
}
\concept{posterior helpers}
