% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_directions}
\alias{wind_directions}
\title{Simulated wind directions}
\format{
A tibble with 500 rows and 4 variables:
\describe{
\item{station}{Station identifier.}
\item{direction}{Wind direction in radians.}
\item{speed}{Wind speed in arbitrary units.}
\item{season}{Season label.}
}
}
\source{
Simulated for package examples.
}
\usage{
wind_directions
}
\description{
A reproducible simulated dataset of wind directions with station, speed and
season variables. Angles are stored in radians.
}
\keyword{datasets}
